resources and community add

+++ PRJ/HopeLine.DataAccess/Entities/Map.cs
@@ -23,10 +23,10 @@ namespace HopeLine.DataAccess.Entities
         //[MaxLength(500)]
         //public ICollection<string> LocationNames { get; set; }
 
-        public float XCoordinate { get; set; }
+        public double XCoordinate { get; set; }
 
-        public float YCoordinate { get; set; }
+        public double YCoordinate { get; set; }
 
-        public float Radius { get; set; }
+        public double Radius { get; set; }
     }
 }
